Rename assets to components
