+++ storybook/src/stories/UseArray/Burger/components/Patty.tsx
